+++ R/fit_svs.R
@@ -7,6 +7,7 @@
 #' @param output_dir directory path to output fitting results.
 #' @param mri filepath or nifti object containing anatomical MRI data.
 #' @param mri_seg filepath or nifti object containing segmented MRI data.
+#' @param deface option to apply fsl_deface to the mri input. Defaults to FALSE.
 #' @param segment_t1 segment the t1 weighted mri file with FSL FAST and use the
 #' results to perform partial volume correction. Defaults to FALSE.
 #' @param external_basis precompiled basis set object to use for analysis.
@@ -116,10 +117,10 @@
 #' }
 #' @export
 fit_svs <- function(input, w_ref = NULL, output_dir = NULL, mri = NULL,
-                    mri_seg = NULL, segment_t1 = FALSE, external_basis = NULL,
-                    append_external_basis = FALSE, p_vols = NULL,
-                    format = NULL, pul_seq = NULL, TE = NULL, TR = NULL,
-                    TE1 = NULL, TE2 = NULL, TE3 = NULL, TM = NULL,
+                    mri_seg = NULL, deface = FALSE, segment_t1 = FALSE,
+                    external_basis = NULL, append_external_basis = FALSE,
+                    p_vols = NULL, format = NULL, pul_seq = NULL, TE = NULL,
+                    TR = NULL, TE1 = NULL, TE2 = NULL, TE3 = NULL, TM = NULL,
                     append_basis = NULL, remove_basis = NULL, pre_align = TRUE,
                     dfp_corr = TRUE, dfp_corr_ref_subset = NULL,
                     output_ratio = NULL, ecc = FALSE, hsvd_width = NULL,
@@ -167,7 +168,8 @@ fit_svs <- function(input, w_ref = NULL, output_dir = NULL, mri = NULL,
       mri_seg <- vector(mode = "list", length = length(input))
     }
     
-    more_args <- list(segment_t1 = segment_t1, external_basis = external_basis,
+    more_args <- list(deface = deface, segment_t1 = segment_t1,
+                      external_basis = external_basis,
                       append_external_basis = append_external_basis,
                       p_vols = p_vols, format = format, pul_seq = pul_seq,
                       TE = TE, TR = TR, TE1 = TE1, TE2 = TE2, TE3 = TE3,
@@ -314,6 +316,14 @@ fit_svs <- function(input, w_ref = NULL, output_dir = NULL, mri = NULL,
     } else {
       mri <- readNifti(mri)
     }
+  }
+  
+  # deface the mri if specified
+  if (is.def(mri) & deface) {
+    dir.create(file.path(output_dir, "mri_deface"), showWarnings = FALSE)
+    deface_path <- file.path(output_dir, "mri_deface", "mri_deface.nii.gz")
+    fslr::fsl_deface(mri, outfile = deface_path, verbose = FALSE)
+    mri <- readNifti(deface_path)
   }
   
   # reorientate mri